Fix earth rotation calculation, which wasn't correctly calculating sidereal time
It seemed to be possibly double-counting sidereal time, in the
calculations for both gmst0 (where there's a 'tu' term that adds 100
days per julian century, or 1 day per year) and gmst, which directly
multiplies by 1.002737909350795 to correct for solar/sidereal time.

I had misinterpreted the calculations at:
https://gssc.esa.int/navipedia/index.php/CEP_to_ITRF
where it talked about UT1 (I assumed UT1 was the UT1 time since midnight
on 1st January 2000, while it was actually the UT1 time since the
previous midnight of whichever day).

The test data I've added was originally checked in Cosmographia, but the
calculations did not quite line up with what Cosmographia showed, so I
optimised the timeUtcMillis to produce the closest rotations to what I
was looking for - this sometimes moved the time by a few minutes. There
is probably a better way of doing this, but I don't necessarily expect
ESA's calculations to line up with JPL's exactly anyway, so I'm leaving
it here for now.