Use 8MB of flash on the ESP32 instead of 4MB
The partition table can only be changed when uploading over a serial
connection, not OTA. huge_app.csv does not include a second app
partition which is required for OTAs.
I could have used min_spiffs.csv instead, but it doesn't give as much
space for the app as this, and my ESP32 has 16MB of flash onboard.
(16MB would also work, but that would be less compatible with other
devices)