Make spacepointer and menu libs independent
This creates a "main" library that just contains things that need to
depend on both libraries.

Before this, all of the spacepointer tests were broken in the native
environment, because these files made spacepointer depend on menu, which
in turn depends on Arduino.h, which doesn't exist in the native
environment.