Improve calibration routine
This stops calibration from crashing if the motor isn't moving the
sensor, by resetting the angle to zero after every movement instead of
adding up the angles and resetting once at the end. When adding up the
angles, the altitude could go out of range and cause an abort.

This also does another altitude check after trying to zero the azimuth,
and goes back to altitude calibration if it has moved away from
horizontal.